// src/pods/character/api/character.api-model.ts
export interface Character {
  id: number;
  name: string;
  bestSentences: string[];
  status: string;
  species: string;
  type: string;
  gender: string;
  origin: {
    name: string;
    url: string;
  };
  location: {
    name: string;
    url: string;
  };
  image: string;
  episode: string[];
  url: string;
  created: string;
}